third_party/blink/web_tests/platform/generic/fast/css/transform-origin-parsing-expected.txt

This tests checks parsing of the '-webkit-transform-origin' property and in particular that specifying invalid z values discards the property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS test("-webkit-transform-origin: 10% 10% 10%", "-webkit-transform-origin") is ""
PASS test("-webkit-transform-origin: 10% 10% 10px", "-webkit-transform-origin") is "10% 10% 10px"
PASS test("-webkit-transform-origin: 10px 10px 10%", "-webkit-transform-origin") is ""
PASS test("-webkit-transform-origin: 10px 10px 10px", "-webkit-transform-origin") is "10px 10px 10px"
PASS test("-webkit-transform-origin: left top 10%", "-webkit-transform-origin") is ""
PASS test("-webkit-transform-origin: left top 10px", "-webkit-transform-origin") is "left top 10px"
PASS test("-webkit-transform-origin: left top left", "-webkit-transform-origin") is ""
PASS successfullyParsed is true

TEST COMPLETE
